% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPD.R
\name{asymptoticCIforGPDfit}
\alias{asymptoticCIforGPDfit}
\title{Title}
\usage{
asymptoticCIforGPDfit(fitGPD, h, hGrad, alpha = 0.05, bonferroni = 1,
  verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
\examples{
library(QRM)
sample <- rGPD(500, 1,1)

# Find a suitable threshold, and fit a GPD
MEplot(sample)
u <- 0
fitGPD <- fit.GPD(sample, u, type = "ml")

# Build a CI on P(4 <= X <= 9), whose true value is 10\%
h <- function(xi, beta) pGPD(9 -u,xi,beta) - pGPD(4 -u,xi,beta)
hGrad <- function(xi,beta) gradientpGPD(9-u,xi,beta) - gradientpGPD(4-u,xi,beta)
asymptoticCIforGPDfit(fitGPD,h,hGrad )
}

