% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{Dexp}
\alias{Dexp}
\title{Derivative w.r.t. x of the exponential distribution function}
\usage{
Dexp(x, D, rate = 1)
}
\arguments{
\item{x}{vector of quantiles}

\item{D}{a non-negative integer}

\item{rate}{of the distributio}
}
\value{
Dexp gives the  distribution function when D = 0, the density
 when D  = 1, the derivative of the density when D = 2, and so on
}
\description{
Derivative w.r.t. x of the exponential distribution function
}
\details{
F(x) = 1 - exp(-rate x)
}
\examples{
Dexp(1,1,0,2)
}

