% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{Dpareto}
\alias{Dpareto}
\title{Title}
\usage{
Dpareto(x, d, scale = 1, shape = 1)
}
\arguments{
\item{shape}{}
}
\description{
Title
}
\examples{
d <- 1:4
x <- seq(-3,15, length = 100)

# Compute the firts 4 derivatives
dataPlot <- NULL
for (D in d)
dataPlot <- rbind(dataPlot, data.frame(d = as.factor(D), x = x, value = (Dpareto(x,D))))

# Plot the derivatives
library(ggplot2)
ggplot(dataPlot, aes(x = x, y = value, color = d)) + geom_line()
}

