% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{partialExpectationlnorm}
\alias{partialExpectationlnorm}
\title{Compute E[X I(X <= x)] when X follows a log-normal distribution}
\usage{
partialExpectationlnorm(x, meanlog = 0, sdlog = 1, lower = TRUE)
}
\arguments{
\item{x}{A positive vector}

\item{meanlog}{mean and standard deviation of the distribution
on the log scale with default values of 0 and 1 respectively.}

\item{sdlog}{mean and standard deviation of the distribution
on the log scale with default values of 0 and 1 respectively.}

\item{lower}{Logical value determining whether the lower or the upper partial
expectation should be computed, the function should return E[X I(X <= x)] or
E[X I(X > x)].  Default is TRUE for lower expectation.}
}
\description{
Compute E[X I(X <= x)] when X follows a log-normal distribution
}
\examples{
partialExpectationlnorm(1, 0,1, lower = TRUE)
}

