% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{ppareto}
\alias{ppareto}
\title{Title}
\usage{
ppareto(x, scale = 1, shape = 1)
}
\arguments{
\item{shape}{}
}
\description{
Title
}
\examples{

x <- seq(-3,15, length = 100)
plot(x,ppareto(x), type ="l")
}

