% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{qpareto}
\alias{qpareto}
\title{Title}
\usage{
qpareto(p, scale = 1, shape = 1)
}
\arguments{
\item{shape}{}
}
\description{
Title
}
\examples{
x <- seq(1,15, length = 100)
p <- ppareto(x)
q <- qpareto(p)
hist(x-q)
}

