% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{rpareto}
\alias{rpareto}
\title{Title}
\usage{
rpareto(n, scale = 1, shape = 1)
}
\arguments{
\item{scale}{}
}
\description{
Title
}
\examples{
# Simulate a pareto sample
n <- 1e3 ; scale = 1/2 ; shape = 3
set.seed(100) ; X <- sort(rpareto(n, scale, shape))

# Compare the ECDF with the theoritical CDF
plot(X,ecdf(X)(X), type = "l")
lines(X,ppareto(X, scale, shape), col = "red")
}

