% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPD.R
\name{asymptoticCIforGPDfit}
\alias{asymptoticCIforGPDfit}
\title{Build (1-alpha) confidence intervals of functions of GPD parameters}
\usage{
asymptoticCIforGPDfit(fitGPD, h, hGrad, alpha = 0.05, verbose = TRUE)
}
\arguments{
\item{fitGPD}{An object obtained using the function \emph{fit.GPD} available in the
package \emph{QRM} available on CRAN}

\item{h}{A function taking for argument the shape \emph{xi}, and scale \emph{beta} parameters
of a GPD distribution}

\item{hGrad}{The gradient of the function h w.r.t to \emph{xi}, and \emph{beta}, in that order}

\item{alpha}{The desired level of accuracy for the confidence interval}

\item{verbose}{A logical value indicating wether messages should be printed on the screen}
}
\value{
a data frame containing three values
\item{lB}{the lower bound of the confidence interval}
\item{hHat}{the estimated value of h(xi,beta)}
\item{uB}{the upper bound of the confidence interval}
}
\description{
This function applies the delta method to derive a (1-alpha) confidence interval
of the function h(xi, beta) where xi, and beta are respectively the MLE estimators
of the shape and scale parameters of a GPD distribution.
}
\examples{
library(QRM)
sample <- rGPD(500, 1,1)

# Find a suitable threshold, and fit a GPD
MEplot(sample)
u <- 0
fitGPD <- fit.GPD(sample, u, type = "ml")

# Build a CI on P(4 <= X <= 9), whose true value is 10\%
h <- function(xi, beta) pGPD(9 -u,xi,beta) - pGPD(4 -u,xi,beta)
hGrad <- function(xi,beta) gradientpGPD(9-u,xi,beta) - gradientpGPD(4-u,xi,beta)
asymptoticCIforGPDfit(fitGPD,h,hGrad )
}

