% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{Dlnorm}
\alias{Dlnorm}
\title{Title}
\usage{
Dlnorm(x, D, meanlog, sdlog)
}
\arguments{
\item{sdlog}{}
}
\description{
Title
}
\examples{
# Simulate a pareto sample
n <- 1e3 ; meanlog <- 1 ; sdlog <- 2
set.seed(100) ; X <- sort(rlnorm(n, meanlog, sdlog))
d <- 1:4
# Compare the ECDF with the theoritical CDF
plot(X,ecdf(X)(X), type = "l")
lines(X,ppareto(X, scale, shape), col = "red")
}

